refactor: 관심사분리
