+++ src/components/PixelCanvas.js
@@ -3,30 +3,14 @@ import React, { useEffect } from 'react';
 import styled from 'styled-components';
 
 import useStore from '../store/store';
+import makeArray from '../utils/makeArray';
 
 import Grid from './Grid';
-
-// 최적화 필요.
-const makeArray = (column, row, color) => {
-  let canvas = Array(column);
-
-  for (let x = 0; x < column; x++) {
-    let canvasRow = Array(column);
-
-    for (let y = 0; y < row; y++) {
-      canvasRow[y] = color;
-    }
-
-    canvas[x] = canvasRow;
-  }
-
-  return canvas;
-};
 
 const PixelCanvas = () => {
   const { canvas, row, column, selectedColor, selectedTools, baseColor } =
     useStore();
-  // 최적화 필요
+
   const update = (x, y) => {
     if (canvas[y][x] !== selectedColor) {
       canvas[y][x] = selectedColor;
